# domain-controller/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이 API는 Ablecloud Works의 Domain Controller(DC)를 제어하는 역할을 합니다.",
        "title": "Ablecloud Works Domain-Controller API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ad-api.ablecloud.io",
    "basePath": "/api/v1",
    "paths": {
        "/app": {
            "get": {
                "description": "윈도우 시작메뉴에 등록된 프로그램의 정보를 출력하는 API",
                "produces": [
                    "application/json"
                ],
                "summary": "윈도우 앱 목록",
                "responses": {
                    "200": {
                        "description": "목록 표시",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.APPVAL"
                            }
                        }
                    },
                    "401": {
                        "description": "명령 실패",
                        "schema": {
                            "$ref": "#/definitions/main.errorModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/cmd/": {
            "get": {
                "description": "powershell 명령 처리기",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "powershell 명령 처리기",
                "parameters": [
                    {
                        "type": "string",
                        "description": "명령어",
                        "name": "cmd",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "인자",
                        "name": "arg",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "시간제한, 기본값",
                        "name": "timeout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "명령 성공",
                        "schema": {
                            "$ref": "#/definitions/main.shellReturnModel"
                        }
                    },
                    "401": {
                        "description": "명령 실패",
                        "schema": {
                            "$ref": "#/definitions/main.errorModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "사용자 로그인 체크",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "사용자 이름",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "사용자 암호",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "로그인 성공",
                        "schema": {
                            "$ref": "#/definitions/main.loginModel"
                        }
                    },
                    "401": {
                        "description": "로그인 실패",
                        "schema": {
                            "$ref": "#/definitions/main.loginModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "사용자 목록 조회",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "사용자 목록 조회",
                "responses": {
                    "200": {
                        "description": "사용자 생성 성공",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ADUser"
                            }
                        }
                    },
                    "401": {
                        "description": "사용자 생성 실패",
                        "schema": {
                            "$ref": "#/definitions/main.errorModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "post": {
                "description": "사용자 생성",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add New User account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "사용자 이름",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "사용자 암호",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "전화번호",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "이메일",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "이름",
                        "name": "givenName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "성",
                        "name": "sn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "직급",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "사용자 생성 성공",
                        "schema": {
                            "$ref": "#/definitions/main.userModel"
                        }
                    },
                    "401": {
                        "description": "사용자 생성 실패",
                        "schema": {
                            "$ref": "#/definitions/main.userModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "description": "사용자 목록 조회",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "사용자 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "사용자 이름",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "사용자 생성 성공",
                        "schema": {
                            "$ref": "#/definitions/main.ADUser"
                        }
                    },
                    "401": {
                        "description": "사용자 생성 실패",
                        "schema": {
                            "$ref": "#/definitions/main.errorModel"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ADUser": {
            "type": "object",
            "properties": {
                "i": {
                    "description": "구/군/시",
                    "type": "string"
                }
            }
        },
        "main.APPVAL": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "main.errorModel": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "main.loginModel": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "login": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "main.shellReturnModel": {
            "type": "object",
            "properties": {
                "stderr": {
                    "type": "string"
                },
                "stdout": {
                    "type": "string"
                }
            }
        },
...
    },
...
}